Change the funcionality of the PostForm from the usePost to PostForm
