+++ src/styles/styles.css
@@ -40,3 +40,7 @@ button {
   cursor: pointer;
   font: inherit;
 }
+
+input {
+  font: inherit;
+}
